local/embedded remote/daemon notification tests
